[ADD] basic fleeing behaviour for Awacs added fleeing from the closest detections in update function added utils get id from self  Changes to be committed: 	modified:   src/ADWG/include/components/Behaviours.hpp 	modified:   src/ADWG/include/components/attributes.hpp 	modified:   src/ADWG/include/components/radar.hpp 	deleted:    src/ADWG/include/components/radar_awacs.hpp 	deleted:    src/ADWG/include/components/radar_fighter.hpp 	deleted:    src/ADWG/include/components/radar_missile.hpp 	modified:   src/ADWG/src/components/Behaviours.cpp 	deleted:    src/ADWG/src/radar_fighter.cpp 	modified:   src/core/system/vector2.inl 	modified:   src/core/utils/angle.cpp 	modified:   src/core/utils/angle.hpp 	new file:   src/core/utils/registry_utils.hpp 	new file:   src/core/utils/registry_utils.inl

+++ src/ADWG/include/components/attributes.hpp
@@ -5,7 +5,7 @@
 
 typedef struct flight_data_s
 {
-    adwg::Vector3<int> position;
+    adwg::Vector3<double> position;
     // default always on 360° basis 
     double orientation; 
 } flight_data_t;
